# docs/swagger.yaml
definitions:
...
  model.JobApplication:
    properties:
      cover_letter:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      deleted_by:
        type: integer
      id:
        type: integer
      is_deleted:
        type: boolean
      notes:
        type: string
      post_job_id:
        type: integer
      resume:
        type: string
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    required:
    - post_job_id
    - resume
    type: object
...
  response.VResponse:
    properties:
      code: {}
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /datn_backend/post-job/{id}/applications:
    get:
      consumes:
      - application/json
      description: Lấy danh sách các ứng viên đã ứng tuyển vào bài đăng của nhà tuyển
        dụng
      parameters:
      - description: Post Job ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.VResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.JobApplication'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.VResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.VResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách ứng viên cho bài đăng
      tags:
      - JobApplication
  /datn_backend/post-job/applications/{id}/status:
    put:
      consumes:
      - application/json
      description: Nhà tuyển dụng cập nhật trạng thái ứng tuyển của ứng viên
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status (pending, reviewing, accepted, rejected)
        in: body
        name: status
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.VResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.VResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.VResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.VResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật trạng thái ứng tuyển
      tags:
      - JobApplication
...
  /datn_backend/post-job/my-applications:
    get:
      consumes:
      - application/json
      description: Lấy danh sách các công việc mà người dùng đã ứng tuyển
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.VResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.JobApplication'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách ứng tuyển của tôi
      tags:
      - JobApplication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"